some updates for tests, extract test util
